+++ tests/test_cli.py
@@ -39,6 +39,7 @@ def test_rebuild_all_models_all_indexes(sheraf_zeo_database):
         [f"{sheraf_zeo_database.uri}&database_name=cli", "rebuild", "tests.test_cli"],
     )
     assert result.exit_code == 0, result.output
+    assert result.output == "", result.output
 
     with sheraf.connection() as conn:
         assert "foo" in conn.root()[CliModel.table]
@@ -66,6 +67,7 @@ def test_rebuild_all_models_one_index(sheraf_zeo_database):
         ],
     )
     assert result.exit_code == 0, result.output
+    assert result.output == "", result.output
 
     with sheraf.connection() as conn:
         assert "foo" in conn.root()[CliModel.table]